+++ pandas/tests/indexes/datetimes/methods/test_normalize.py
@@ -1,4 +1,8 @@
+from dateutil.tz import tzlocal
 import numpy as np
+import pytest
+
+import pandas.util._test_decorators as td
 
 from pandas import (
     DatetimeIndex,
@@ -37,3 +41,55 @@ class TestNormalize:
         result = dti.normalize()
         expected = DatetimeIndex([NaT, Timestamp("2018-01-01")])
         tm.assert_index_equal(result, expected)
+
+    def test_normalize_tz(self):
+        rng = date_range("1/1/2000 9:30", periods=10, freq="D", tz="US/Eastern")
+
+        result = rng.normalize()  # does not preserve freq
+        expected = date_range("1/1/2000", periods=10, freq="D", tz="US/Eastern")
+        tm.assert_index_equal(result, expected._with_freq(None))
+
+        assert result.is_normalized
+        assert not rng.is_normalized
+
+        rng = date_range("1/1/2000 9:30", periods=10, freq="D", tz="UTC")
+
+        result = rng.normalize()
+        expected = date_range("1/1/2000", periods=10, freq="D", tz="UTC")
+        tm.assert_index_equal(result, expected)
+
+        assert result.is_normalized
+        assert not rng.is_normalized
+
+        rng = date_range("1/1/2000 9:30", periods=10, freq="D", tz=tzlocal())
+        result = rng.normalize()  # does not preserve freq
+        expected = date_range("1/1/2000", periods=10, freq="D", tz=tzlocal())
+        tm.assert_index_equal(result, expected._with_freq(None))
+
+        assert result.is_normalized
+        assert not rng.is_normalized
+
+    @td.skip_if_windows
+    @pytest.mark.parametrize(
+        "timezone",
+        [
+            "US/Pacific",
+            "US/Eastern",
+            "UTC",
+            "Asia/Kolkata",
+            "Asia/Shanghai",
+            "Australia/Canberra",
+        ],
+    )
+    def test_normalize_tz_local(self, timezone):
+        # GH#13459
+        with tm.set_timezone(timezone):
+            rng = date_range("1/1/2000 9:30", periods=10, freq="D", tz=tzlocal())
+
+            result = rng.normalize()
+            expected = date_range("1/1/2000", periods=10, freq="D", tz=tzlocal())
+            expected = expected._with_freq(None)
+            tm.assert_index_equal(result, expected)
+
+            assert result.is_normalized
+            assert not rng.is_normalized
